requirements.txt
asgiref==3.2.7
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.10
django-allauth==0.39.1
django-ckeditor==5.9.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-import-export==2.0.2
django-js-asset==1.2.2
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.0.1
odfpy==1.4.1
openpyxl==3.0.3
pep8==1.7.1
Pillow==6.2.0
pycodestyle==2.5.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
PyYAML==5.3.1
requests==2.21.0
requests-oauthlib==1.2.0
sorl-thumbnail==12.6.3
sqlparse==0.2.4
stripe==2.27.0
tablib==1.1.0
urllib3==1.24.2
xlrd==1.2.0
xlwt==1.3.0
django-cors-headers 
PyMySQL==0.9.2
gunicorn==20.0.4
django-heroku
boto3==1.11.12
django-storages==1.9.1